Fixed some formatting with presenters
