+++ src/Main/browseFunctions/SinglePostPresenterInterface.java
@@ -4,4 +4,5 @@ import postFunctions.Post;
 
 public interface SinglePostPresenterInterface {
     public void presentSinglePost(Post post);
+    public void presentProduct();
 }
